플레이리스트 상세보기 공통 컴포넌트 생성
